Prevent misfire on non-game pages
We use the presence of a `.main-board` element to determine that we're
in a game on the site, and then watch for mutations of the `.material-top`
element to determine if and when we lose our Queen. Some pages, namely
those in the Tools section, have the former but not the latter, which
caused us to try to observe mutations on an element that would never exist.