Refactor to track game state as moves are played
We want to be able to resign as fast as possible to maximize the element
of surprise. Having to parse the DOM into a `Chess` instance when our
Queen gets taken is "fast enough", but it can be faster.

Instead of watching the material imbalance to determine that our Queen
has been taken, we simply apply moves as they're made; chess.js lets us
know when a move resulted in a piece being captured, and we can check to
see whether it's our clock that's running after a Queen capture in order
to begin checking whether it's time to resign.